fix: Add `pad` field in `CubedNode` interface.
The production build cannot figure out the classes' constructor
names. Because of this, the switch-case statement we use to narrow down
the types based on the constructor name won't work. To avoid this, when
creating a new node the first time, we pass the `pad` to the node in
order to set the current pad everytime we need it.